LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/json-module/valid-content-type-expected.txt

CONSOLE MESSAGE: line 1: TypeError: 'text/json' is not a valid JavaScript MIME type.
CONSOLE MESSAGE: line 1: TypeError: 'application/json' is not a valid JavaScript MIME type.
CONSOLE MESSAGE: line 1: TypeError: 'text/html json' is not a valid JavaScript MIME type.
CONSOLE MESSAGE: line 1: TypeError: 'image/svg json' is not a valid JavaScript MIME type.

FAIL text/json assert_unreached: [object Event] Reached unreachable code
FAIL application/json assert_unreached: [object Event] Reached unreachable code
FAIL text/html+json assert_unreached: [object Event] Reached unreachable code
FAIL image/svg+json assert_unreached: [object Event] Reached unreachable code
